+++ internal/dto/admin/management.go
@@ -23,7 +23,7 @@ type DashboardResponse struct {
 
 // BatchVideoRequest is the batch video operation payload.
 type BatchVideoRequest struct {
-	IDs    []uint64 `json:"ids" validate:"required,min=1,max=500"`
+	IDs    []uint32 `json:"ids" validate:"required,min=1,max=500"`
 	Status *uint8   `json:"status" validate:"omitempty,oneof=0 1"`
 }
 
@@ -39,16 +39,16 @@ type AdminListRequest struct {
 	dto.PaginationRequest
 	Keyword string  `form:"keyword"`
 	Status  *uint8  `form:"status"`
-	GroupID *uint64 `form:"group_id"`
+	GroupID *uint32 `form:"group_id"`
 }
 
 // AdminItem is the admin list item.
 type AdminItem struct {
-	ID          uint64 `json:"id"`
+	ID          uint32 `json:"id"`
 	Username    string `json:"username"`
 	Email       string `json:"email"`
 	Avatar      string `json:"avatar"`
-	GroupID     uint64 `json:"group_id"`
+	GroupID     uint32 `json:"group_id"`
 	GroupName   string `json:"group_name"`
 	Status      uint8  `json:"status"`
 	LastLoginAt string `json:"last_login_at"`
@@ -61,7 +61,7 @@ type CreateAdminRequest struct {
 	Password string `json:"password" validate:"required,min=6,max=72"`
 	Email    string `json:"email" validate:"omitempty,email,max=128"`
 	Avatar   string `json:"avatar" validate:"omitempty,max=500"`
-	GroupID  uint64 `json:"group_id" validate:"required,min=1"`
+	GroupID  uint32 `json:"group_id" validate:"required,min=1"`
 	Status   *uint8 `json:"status" validate:"omitempty,oneof=0 1"`
 }
 
@@ -70,7 +70,7 @@ type UpdateAdminRequest struct {
 	Username string  `json:"username" validate:"omitempty,min=3,max=50"`
 	Email    *string `json:"email" validate:"omitempty,max=128"`
 	Avatar   string  `json:"avatar" validate:"omitempty,max=500"`
-	GroupID  *uint64 `json:"group_id" validate:"omitempty,min=1"`
+	GroupID  *uint32 `json:"group_id" validate:"omitempty,min=1"`
 	Status   *uint8  `json:"status" validate:"omitempty,oneof=0 1"`
 }
 
@@ -89,7 +89,7 @@ type UserGroupListRequest struct {
 
 // UserGroupItem is the user group list item.
 type UserGroupItem struct {
-	ID          uint64  `json:"id"`
+	ID          uint32  `json:"id"`
 	Name        string  `json:"name"`
 	Permissions *string `json:"permissions"`
 	Description string  `json:"description"`
@@ -121,7 +121,7 @@ type UserListRequest struct {
 
 // UserItem is the user list item.
 type UserItem struct {
-	ID          uint64 `json:"id"`
+	ID          uint32 `json:"id"`
 	StrID       string `json:"str_id"`
 	Username    string `json:"username"`
 	Nickname    string `json:"nickname"`
@@ -160,11 +160,11 @@ type ResetUserPasswordRequest struct {
 
 // BannerItem is the banner list item.
 type BannerItem struct {
-	ID      uint64 `json:"id"`
+	ID      uint32 `json:"id"`
 	Title   string `json:"title"`
 	Cover   string `json:"cover"`
 	Link    string `json:"link"`
-	VideoID uint64 `json:"video_id"`
+	VideoID uint32 `json:"video_id"`
 	Sort    uint32 `json:"sort"`
 	Status  uint8  `json:"status"`
 }
@@ -174,7 +174,7 @@ type CreateBannerRequest struct {
 	Title   string `json:"title" validate:"required,min=1,max=128"`
 	Cover   string `json:"cover" validate:"required,max=500"`
 	Link    string `json:"link" validate:"omitempty,max=500"`
-	VideoID uint64 `json:"video_id" validate:"omitempty,min=1"`
+	VideoID uint32 `json:"video_id" validate:"omitempty,min=1"`
 	Sort    uint32 `json:"sort" validate:"omitempty,min=0"`
 	Status  *uint8 `json:"status" validate:"omitempty,oneof=0 1"`
 }
@@ -184,7 +184,7 @@ type UpdateBannerRequest struct {
 	Title   string  `json:"title" validate:"omitempty,min=1,max=128"`
 	Cover   string  `json:"cover" validate:"omitempty,max=500"`
 	Link    string  `json:"link" validate:"omitempty,max=500"`
-	VideoID *uint64 `json:"video_id" validate:"omitempty,min=1"`
+	VideoID *uint32 `json:"video_id" validate:"omitempty,min=1"`
 	Sort    *uint32 `json:"sort" validate:"omitempty,min=0"`
 	Status  *uint8  `json:"status" validate:"omitempty,oneof=0 1"`
 }
