+++ internal/dto/open/category.go
@@ -2,7 +2,7 @@ package open
 
 // CategoryItem is a flat category payload for the open API.
 type CategoryItem struct {
-	ID       uint64 `json:"id"`
+	ID       uint32 `json:"id"`
 	Name     string `json:"name"`
-	ParentID uint64 `json:"parent_id"`
+	ParentID uint32 `json:"parent_id"`
 }
